More using statements for Dynamo,Logging. And add LogMessage.
