+++ src/DynamoCore/Extensions/ExtensionManager.cs
@@ -1,4 +1,6 @@
 ﻿using Dynamo.Interfaces;
+using Dynamo.Logging;
+
 using System;
 using System.Collections.Generic;
 using System.Linq;
