+++ src/DynamoCoreWpf/Extensions/ViewExtensionLoader.cs
@@ -1,4 +1,5 @@
 ﻿using Dynamo.Interfaces;
+using Dynamo.Logging;
 
 using System;
 using System.Collections.Generic;
