+++ src/DynamoCoreWpf/Views/Core/DynamoView.xaml.cs
@@ -32,6 +32,7 @@ using System.Windows.Controls.Primitives;
 using System.Windows.Media;
 using Dynamo.Services;
 using Dynamo.Wpf.Utilities;
+using Dynamo.Logging;
 
 using ResourceNames = Dynamo.Wpf.Interfaces.ResourceNames;
 using Dynamo.Wpf.ViewModels.Core;
